# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "OXO Game API",
        "contact": {
            "name": "Michael Sum",
            "email": "sumchuman@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/challenges": {
            "post": {
                "description": "Player joins a challenge by POST, returns status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "Join a challenge",
                "parameters": [
                    {
                        "description": "Challenge information",
                        "name": "challenge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_internal_models.Challenge"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.JoinResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/challenges/results": {
            "get": {
                "description": "Fetches all challenge results",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "challenges"
                ],
                "summary": "Get challenge results",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/oxo-game-api_internal_models.ChallengeResult"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/levels": {
            "get": {
                "description": "Get all levels with details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "levels"
                ],
                "summary": "Get",
                "parameters": [
                    {
                        "description": "Level information",
                        "name": "level",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_internal_models.Level"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_internal_models.Level"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new level with the specified name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "levels"
                ],
                "summary": "Create a new level",
                "parameters": [
                    {
                        "description": "Level information",
                        "name": "level",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_internal_models.Level"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.LevelCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/logs": {
            "get": {
                "description": "Fetches game logs based on query parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Get game logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "player_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Action type",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time in RFC3339 format",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time in RFC3339 format",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit the number of logs",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/oxo-game-api_internal_models.GameLog"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new game log entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Create a new game log",
                "parameters": [
                    {
                        "description": "Game log information",
                        "name": "log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_internal_models.GameLog"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.LogCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/oxo-game-api_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "oxo-game-api_internal_models.Challenge": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "player_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "oxo-game-api_internal_models.ChallengeResult": {
            "type": "object",
            "properties": {
                "challenge": {
                    "$ref": "#/definitions/oxo-game-api_internal_models.Challenge"
                },
                "challenge_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "player": {
                    "$ref": "#/definitions/oxo-game-api_internal_models.Player"
                },
                "player_id": {
                    "type": "integer"
                },
                "won": {
                    "type": "boolean"
                }
            }
        },
        "oxo-game-api_internal_models.GameLog": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "player_id": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "oxo-game-api_internal_models.Level": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "max_exp": {
                    "type": "integer"
                },
                "min_exp": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "oxo-game-api_internal_models.Player": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "$ref": "#/definitions/oxo-game-api_internal_models.Level"
                },
                "level_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "oxo-game-api_pkg_utils_response.JoinResponse": {
            "type": "object",
            "properties": {
                "challenge_id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success_join": {
                    "type": "boolean"
                }
            }
        },
        "oxo-game-api_pkg_utils_response.LevelCreateResponse": {
            "type": "object",
            "properties": {
                "level_id": {
                    "type": "integer"
                }
            }
        },
        "oxo-game-api_pkg_utils_response.LogCreateResponse": {
            "type": "object",
            "properties": {
                "log_id": {
                    "type": "integer"
                }
            }
        },
...
        "oxo-game-api_pkg_utils_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}